+++ CHANGELOG.md
@@ -9,11 +9,28 @@ unreleased
 
 - public version info placed in version.h
 
+- automatic build (via autotools) of some example programs (issue #3)
+
 ## Removed
 
 - fetch and link of googletest, was not being used
 
+## Changed
+
+- wamp_session::provide() accept callback that is invoked on success/failure of
+  registration.
+
+- Using msgpack-c 2.1.2.  In this version the header-only msgpack decoder has
+  been fixed, which is used by wampcc.
+
+- The stream logger (used by the logger::console logger) now takes a wrapper to
+  a ostream& and a mutex, instead of just an ostream reference.  This is so that
+  a synchronization mechanism is available to synchronize writes to the
+  stream (issue #2, petten).
+
 ## Fixed
+
+- wamp_session not handling failed registration
 
 - compile errors on Xcode 7.3.1 + 10.7 SDK (issue #2, petten)